move document context token
