+++ src/packages/documents/documents/workspace/document-workspace.context.ts
@@ -270,12 +270,3 @@ export class UmbDocumentWorkspaceContext
 }
 
 export default UmbDocumentWorkspaceContext;
-
-export const UMB_DOCUMENT_WORKSPACE_CONTEXT = new UmbContextToken<
-	UmbSaveableWorkspaceContextInterface,
-	UmbDocumentWorkspaceContext
->(
-	'UmbWorkspaceContext',
-	undefined,
-	(context): context is UmbDocumentWorkspaceContext => context.getEntityType?.() === UMB_DOCUMENT_ENTITY_TYPE,
-);
